requirements.txt
# ReplenishIQ Dependencies
# Python 3.10+

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Database connections
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
snowflake-connector-python>=3.0.0

# Machine learning and forecasting
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter notebooks
jupyter>=1.0.0
notebook>=7.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
tqdm>=4.65.0
